# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "xg backend rest api",
        "title": "xg REST API",
        "termsOfService": "https://localhost:8088/v1",
        "contact": {},
        "version": "0.1 alpha"
    },
    "paths": {
...
        "/api/order_source": {
            "post": {
                "description": "create an order source",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orderSource"
                ],
                "summary": "createOrderSource",
                "parameters": [
                    {
                        "description": "create order source",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.CreateOrderSourceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
        "/api/order_sources": {
            "get": {
                "description": "list all order sources",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orderSource"
                ],
                "summary": "listOrderSources",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.OrderSource"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.CreateOrderSourceRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "entity.OrderSource": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "route.Response": {
            "type": "object",
            "properties": {
                "err_msg": {
                    "type": "string"
                }
            }
        },
...
    }
}